# openapi.yaml
openapi: 3.1.0
info:
  title: HTML Publisher API
  description: |
    HTMLコンテンツをGitHub Gistに公開し、ツールとして共有するためのAPI。

    ## 基本的な使い方

    ### 1. 新規作成
    `POST /api/tools` でHTMLを送信すると、Gistに保存されIDとURLが返されます。

    ### 2. 閲覧
    - **ブラウザで表示**: 返されたURL（例: `https://example.com/tool/abc123`）にアクセス
    - **HTMLソースを取得**: `GET /api/tools/{id}` でHTMLコンテンツをJSON形式で取得

    **注意**: `/tool/{id}` はSPAページのため、APIからは直接読み取れません。
    HTMLの中身を確認・取得するには必ず `GET /api/tools/{id}` を使用してください。

    ### 3. 更新
    既存のツールを更新するには `PUT /api/tools/{id}` を使用します。
    IDはURLの末尾部分です（例: URLが `https://example.com/tool/abc123` の場合、IDは `abc123`）。

    ### IDの特定方法
    - 作成時のレスポンスに含まれる `id` フィールド
    - ツールURL `/tool/{id}` の `{id}` 部分
    - 例: `https://example.com/tool/abc123def456` → ID: `abc123def456`
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT

servers:
  - url: https://example.com
    description: 本番環境
  - url: http://localhost:3000
    description: ローカル開発環境

paths:
  /api/tools:
    get:
      summary: ツール一覧を取得
      description: |
        直近で作成・更新されたツールの一覧を取得します（最大10件）。
        HTMLソースは含まれません。

        **永続モード（Gist）のエントリのみ返されます**。
        揮発モード（cache、`c_` で始まるID）はプライバシー保護のため一覧に含まれません。
        揮発モードのツールは作成時に返される URL/ID を記録し、共有相手にのみ伝える運用を想定しています。
      operationId: listTools
      tags:
        - Tools
      parameters:
        - name: limit
          in: query
          required: false
          description: 取得件数（1-10、デフォルト10）
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 10
          example: 5
      responses:
        "200":
          description: ツール一覧が正常に取得されました
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolSummary"
              example:
                - id: "abc123def456"
                  name: "コンパスアプリ"
                  memo: "方位を表示"
                  trust: true
                  url: "https://your-domain.com/tool-trust/abc123def456"
                  updatedAt: "2024-01-15T12:00:00Z"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
    post:
      summary: 新しいツールを作成
      description: |
        HTMLコンテンツをアップロードし、公開URLを生成します。

        レスポンスに含まれる `id` と `url` を保存してください。
        - `url`: ブラウザで開くとHTMLが表示されるURL（trustの値によって `/tool/` または `/tool-trust/` になります）
        - `id`: 後で更新する際に使用するID（URLの末尾部分と同じ）

        **認証と保存先**:
        - APIキー認証あり → `ephemeral` の指定に従い、Gist (永続) または cache (揮発) に保存
        - APIキー未提示 → **揮発モード固定** (cache 保存、最終アクセスから約5時間で自動削除)
      operationId: createTool
      tags:
        - Tools
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolRequest"
            example:
              html: "<html><body><h1>Hello World</h1></body></html>"
              name: "サンプルツール"
              memo: "初回作成"
              trust: false
      responses:
        "201":
          description: ツールが正常に作成されました
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateToolResponse"
              example:
                id: "abc123def456"
                url: "https://your-domain.com/tool/abc123def456"
                rawUrl: "https://gist.githubusercontent.com/..."
                trust: false
        "400":
          description: リクエストが不正です
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "html is required and must be a string"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"

  /api/openapi:
    get:
      summary: OpenAPI仕様を取得
      description: このAPIのOpenAPI仕様をYAMLまたはJSON形式で取得します
      operationId: getOpenApiSpec
      tags:
        - Documentation
      parameters:
        - name: format
          in: query
          required: false
          description: 出力形式（yaml または json）
          schema:
            type: string
            enum:
              - yaml
              - json
            default: yaml
          example: "json"
      responses:
        "200":
          description: OpenAPI仕様
          content:
            text/yaml:
              schema:
                type: string
            application/json:
              schema:
                type: object

  /api/tools/{id}:
    get:
      summary: ツールのHTMLソースを取得
      description: |
        指定されたIDのツールのHTMLコンテンツを取得します。

        **重要**: `/tool/{id}` または `/tool-trust/{id}` はブラウザ表示用のSPAページです。
        HTMLの中身を確認・取得するには、このAPIエンドポイント `GET /api/tools/{id}` を使用してください。

        レスポンスの `html` フィールドに完全なHTMLソースが含まれます。
      operationId: getTool
      tags:
        - Tools
      parameters:
        - name: id
          in: path
          required: true
          description: |
            ツールのID。URLの `/tool/` または `/tool-trust/` の後ろの部分です。
            例: URLが `https://example.com/tool/abc123def456` の場合、IDは `abc123def456`
          schema:
            type: string
          example: "abc123def456"
      responses:
        "200":
          description: ツールが正常に取得されました
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetToolResponse"
              example:
                id: "abc123def456"
                html: "<html><body><h1>Hello World</h1></body></html>"
                rawUrl: "https://gist.githubusercontent.com/..."
                url: "https://your-domain.com/tool/abc123def456"
                trust: false
        "400":
          description: リクエストが不正です
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "id is required"
        "404":
          description: ツールが見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Tool not found"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"
    put:
      summary: ツールを更新（HTML上書き）
      description: |
        指定されたIDのツールのHTMLコンテンツを上書き更新します。

        **IDの取得方法:**
        - 作成時のレスポンスに含まれる `id` フィールド
        - ツールURL `/tool/{id}` または `/tool-trust/{id}` から抽出（例: `https://example.com/tool/abc123` → `abc123`）

        `trust` フラグを変更すると、URLのパスが `/tool/` から `/tool-trust/` に変わります（またはその逆）。

        **認証**:
        - APIキー認証あり → 永続(Gist)・揮発(cache) どちらの ID も更新可能
        - APIキー未提示 → 揮発モード（`c_` で始まるID）のみ更新可能。永続Gistの更新は 403 エラー
      operationId: updateTool
      tags:
        - Tools
      parameters:
        - name: id
          in: path
          required: true
          description: |
            ツールのID。URLの `/tool/` または `/tool-trust/` の後ろの部分です。
            例: URLが `https://example.com/tool/abc123def456` の場合、IDは `abc123def456`
          schema:
            type: string
          example: "abc123def456"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolRequest"
            example:
              html: "<html><body><h1>Updated Content</h1></body></html>"
              name: "更新されたツール"
              memo: "ボタンの色を変更"
              trust: true
      responses:
        "200":
          description: ツールが正常に更新されました
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateToolResponse"
              example:
                id: "abc123def456"
                url: "https://your-domain.com/tool-trust/abc123def456"
                rawUrl: "https://gist.githubusercontent.com/..."
                trust: true
        "400":
          description: リクエストが不正です
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "html is required and must be a string"
        "403":
          description: 認証なしで永続Gistの更新を試みた
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "永続モード（Gist）の更新には認証が必要です。揮発モード（c_で始まるID）のみ匿名で更新可能です"
        "404":
          description: ツールが見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Tool not found"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Internal server error"

components:
  schemas:
    CreateToolRequest:
      type: object
      required:
        - html
      properties:
        html:
          type: string
          description: 公開するHTMLコンテンツ
          minLength: 1
        name:
          type: string
          description: ツール名（任意）。Gist説明とHTML内metaタグに反映されます
        memo:
          type: string
          description: 変更内容のメモ（任意）。Gist説明とHTML内metaタグに反映されます
        trust:
          type: boolean
          description: |
            信頼モード（任意）。trueの場合はlocalStorage等が使える信頼済みURLを発行します。
            URLは `/tool-trust/{id}` になります。
        ephemeral:
          type: boolean
          description: |
            揮発モード（任意）。trueの場合はGistではなくキャッシュに保存され、
            最後のアクセスから一定時間（デフォルト6時間）で自動削除されます。
            ハッカソンや一時共有用途に推奨。GITHUB_TOKEN未設定時は自動的に揮発モードになります。
            揮発モードのIDは `c_` プレフィックス付き。

    CreateToolResponse:
      type: object
      required:
        - id
        - url
      properties:
        id:
          type: string
          description: ツールのID（揮発モードは `c_` プレフィックス付き、それ以外はGist ID）
        url:
          type: string
          format: uri
          description: ツールの公開URL（trustの値によって `/tool/` または `/tool-trust/` になります）
        rawUrl:
          type: string
          format: uri
          description: GistのRaw URL（永続モードのみ）
        trust:
          type: boolean
          description: 信頼モードが有効かどうか
        mode:
          type: string
          enum: [gist, cache]
          description: ストレージモード（gist=永続、cache=揮発）

    GetToolResponse:
      type: object
      required:
        - id
        - html
        - url
      properties:
        id:
          type: string
          description: ツールのID
        html:
          type: string
          description: HTMLコンテンツ
        rawUrl:
          type: string
          format: uri
          description: GistのRaw URL（永続モードのみ）
        url:
          type: string
          format: uri
          description: ツールの公開URL
        trust:
          type: boolean
          description: 信頼モードが有効かどうか
        mode:
          type: string
          enum: [gist, cache]
          description: ストレージモード（gist=永続、cache=揮発）

    ToolSummary:
      type: object
      required:
        - id
        - url
        - updatedAt
      properties:
        id:
          type: string
          description: ツールのID
        name:
          type: string
          description: ツール名
        memo:
          type: string
          description: 変更メモ
        trust:
          type: boolean
          description: 信頼モードが有効かどうか
        url:
          type: string
          format: uri
          description: ツールの公開URL
        updatedAt:
          type: string
          format: date-time
          description: 最終更新日時
        mode:
          type: string
          enum: [gist, cache]
          description: ストレージモード（gist=永続、cache=揮発）

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: エラーメッセージ

tags:
  - name: Tools
    description: |
      HTMLツールの作成・取得・更新・一覧。

      **ワークフロー:**
      1. `GET /api/tools` で既存ツール一覧を確認
      2. `POST /api/tools` で新規作成 → IDとURLを取得
      3. URLにアクセスしてHTMLを確認（通常: `/tool/{id}`、信頼モード: `/tool-trust/{id}`）
      4. 修正が必要な場合は `PUT /api/tools/{id}` で上書き更新

      **信頼モード（trust）:**
      - `trust: true` で作成すると `/tool-trust/{id}` のURLが発行され、localStorage等が使用可能
      - `trust: false` または未指定の場合は `/tool/{id}` のURLが発行され、セキュリティ制限が厳格
  - name: Documentation
    description: APIドキュメント